Refactor photo/db
